requirements.txt
Appium_Python_Client==2.6.2
pytest==7.1.3
requests==2.28.1
selenium==4.4.3
webdriver_manager==3.8.3
